backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
PyGithub==2.1.1
azure-identity==1.15.0
azure-ai-textanalytics==5.3.0
python-jose==3.3.0
flask-cors==4.0.0
gunicorn==21.2.0
openai==1.12.0 